+++ src/rcpp-rsqlite.cpp
@@ -1,0 +1,16 @@
+#include <Rcpp.h>
+#include <RSQLite.h>
+using namespace Rcpp;
+
+//' RSQLite version
+//' 
+//' @return A character vector containing header and library versions of 
+//'   RSQLite.
+//' @export
+// [[Rcpp::export]]
+CharacterVector rsqliteVersion() {
+  return CharacterVector::create(
+    _["header"] = SQLITE_VERSION,
+    _["library"] = sqlite3_libversion()
+  );
+}